feat: add "sahko" command
The command reads the open API for electrical prices in Finland from
porssisahko.net. The prices are then rendered on a chart, which is sent
to the chat as an image in Telegram. The chart is rendered in-memory,
and cached to help with spam.

I suppose one could take the caching to a higher level also, by
resending the image in Telegram with the image's ID, instead of
uploading a new image. But that would require actually knowing what I am
doing.

The current implementation still has some TODO:s, as there are some
cases with error handling that should probably be improved, and the
whole infrastructure of how images are sent in TG should probably be
refactored/improved.

See: https://porssisahko.net/api