fix: parse ranks correctly (#406)
We are getting the "skill level":th element from the list to show the
rank. If the Leetify API returns a 0, it means the player is unranked.
A Skill Level of 4 would mean the player is a Silver IV, which means the
fourth element should be returned.

Before this change, there would be an off by one error that made the
ranks show the wrong one. By adding the unranked text as well to the
list, we can avoid having to deal with "-1" in places, and have a 1-1
mapping of list index->rank.