Support 15 minute electricity pricing
We switch from the v1 to v2 API, which returns a list of prices per 15
minutes instead of the previously used 1 hour interval. The API is
already returning the 15 minute times in preparation for the upcoming
change to this system.

As we are drawing the bar chart, we also need to update the "width" of
each column, as we are currently drawing "1 hour" forward, we now need
to draw 15 minutes forward.

There might still be an edge case (hehe) with how the initial hour is
handled and how the chart will look at different times, as the previous
APIs first element in the returned data was always an even hour, but now
it can be 00, 15, 30 or 45 minutes past. Based on quick testing it does
not look different though.