fix: Align chart by removing data instead of adding bogus
Previously we added one fake hour of data to align the chart so that it
started at midnight Finnish time. This implementation broke with the
change to 15-minute intervals.

Changed the approach to remove 3 hours of data instead, so that the
chart now will start from 04:00 instead of midnight. This way we only
show real data, and it looks good enough :tm: anyways, although it does
not start at midnight.