feat: Add colors to electricity price chart
Using colorous to generate a gradient of colors for the "bars" that will
hopefully make it more easy to see what price is when, as you can trail
the color all the way to the x-axis.

The gradients in colorous were not the best, so we picked one, and
implemented a "nice" system with Tulttan to essentially reverse the
gradient, and then "skip" a part of the first colors of it that we did
not want included.

Also changed the approach from generating a line to generating a bunch
of rectangles instead, as that seemed like one of the easiest ways of
introducing these kind of colors.